fix(preprocessing): remove competitive behaviour between `ExperimentVideo.open_video` and `ExperimentVideo.data.setter` when defining the end index
The competition happened because setting video data and opening video cancelled the end index definition of one another. To fix this, we define the default end index as `None`, meaning "until the end". Now opening the video only sets the end index when this information hasn't come yet from the video data.