fix(preprocessing): allow omitting `frames_tensor_path` when converting `ExperimentVideo`s to datasets without saving
Previously, if no frames tensor path was defined, `ev.frames_to_tensor()` would fail even if `save==False` and `overwrite==False`. However, in this condition, there is no communication with the file system, and hence there is no need for `frames_tensor_path` to be defined. This modification solves this bug.