feat(datasets/sliceable): define sliceable datasets based on slicerators and almost compatible with TensorFlow datasets
Two main classes are defined: a general `SliceableDataset`, which resembles a `tf.data.Dataset` capable of dealing with slices, multiple indices and boolean masks (operating with `slicerator.Slicerator` under the hood). Not all TensorFlow functionality is available yet here. The second class is `SupervisedSliceableDataset` which is a specialized version for pairs `(x, y)` where `x` is a *feature vector* and `y` is its associated *target*.