perf(io/json): make `json.serialize` a regular function
Previously, `json.serialize` was a `classes.typeclass` instance.
This works very well, but may get very slow, especially when this
is done in hotloops of the program.
Instead, `json.serialize` is not expected to specialize for
different types of inputs. Instead, this is a responsibility for
`json.encode`. Hence, `json.serialize` can be greatly simplified
and converted to a regular function.