refactor(automl/tuning): do not persist tune model returns
The autotuning algorithm is naturally skipped if it has already been executed.
However, if we persist the results, then we cannot guarantee that the hypermodel is correctly initialized, which causes many problems around re-loading stored models.
By skipping the persistence step, hypermodels are always correctly initialized, which allows us to use its interface as expected.